requirements.txt
beautifulsoup4==4.11.1
discord.py==2.0.1
requests==2.22.0
setuptools==45.2.0
